+++ homework/matrix.cpp
@@ -29,7 +29,7 @@ namespace pc {
     float convertTo<float>(const std::string &str) {
         if(str == "i") return std::numeric_limits<float>::infinity();
         if(str == "inf") return std::numeric_limits<float>::infinity();
-        return std::stod(str);
+        return std::stof(str);
     }
 
     template <>
